requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
pyasn1==0.4.5
pysmb==1.1.27
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1
